+++ src/validation/multi_visitor.rs
@@ -102,10 +102,10 @@ impl<'a> Visitor<'a> for MultiVisitor<'a> {
         self.visit_all(|v| v.exit_null_value(ctx, n.clone()));
     }
 
-    fn enter_int_value(&mut self, ctx: &mut ValidatorContext<'a>, i: Spanning<i64>) {
+    fn enter_int_value(&mut self, ctx: &mut ValidatorContext<'a>, i: Spanning<i32>) {
         self.visit_all(|v| v.enter_int_value(ctx, i.clone()));
     }
-    fn exit_int_value(&mut self, ctx: &mut ValidatorContext<'a>, i: Spanning<i64>) {
+    fn exit_int_value(&mut self, ctx: &mut ValidatorContext<'a>, i: Spanning<i32>) {
         self.visit_all(|v| v.exit_int_value(ctx, i.clone()));
     }
